refactor: extract useTimelineScroll hook
